shop_app1/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

jinja2==3.1.4
python-multipart==0.0.9

SQLAlchemy==2.0.35
pydantic==2.9.2
python-dotenv==1.0.1
reportlab==4.2.2
openpyxl==3.1.5

annotated-types==0.7.0
anyio==4.10.0
bcrypt==4.3.0
certifi==2025.8.3
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.115.0
greenlet==3.2.4
h11==0.16.0
httptools==0.6.4
idna==3.10
itsdangerous==2.2.0
MarkupSafe==3.0.2
openpyxl==3.1.5
passlib==1.7.4
pillow==11.3.0
pip==22.2.2
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
PyYAML==6.0.2
reportlab==4.2.2
requests==2.32.5
setuptools==63.2.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.30.6
watchfiles==1.1.0
websockets==15.0.1
